Revert "test: show that the old instructions still work"
This reverts commit 59e920aac6f3e3d2c5aab1afc2ae95ccb9a641e9.
