+++ test/app/android/app/src/main/java/com/example/twiliovoicereactnative/MainApplication.java
@@ -2,22 +2,41 @@ package com.example.twiliovoicereactnative;
 
 import android.app.Application;
 import android.content.Context;
+
+import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactInstanceManager;
+import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
-import java.lang.reflect.InvocationTargetException;
 
+import java.lang.reflect.InvocationTargetException;
+import java.util.List;
+
+import com.twiliovoicereactnative.TwilioVoiceReactNativePackage;
 import com.twiliovoicereactnative.VoiceApplicationProxy;
 
 public class MainApplication extends Application implements ReactApplication {
-  private final VoiceApplicationProxy voiceApplicationProxy;
-  private final MainReactNativeHost mReactNativeHost;
-  public MainApplication() {
-    super();
-    mReactNativeHost = new MainReactNativeHost(this);
-    voiceApplicationProxy = new VoiceApplicationProxy(mReactNativeHost);
-  }
+  private final VoiceApplicationProxy voiceApplicationProxy = new VoiceApplicationProxy(this);
+
+  private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+    @Override
+    public boolean getUseDeveloperSupport() {
+      return BuildConfig.DEBUG;
+    }
+
+    @Override
+    protected List<ReactPackage> getPackages() {
+      List<ReactPackage> packages = new PackageList(this).getPackages();
+      return packages;
+    }
+
+    @Override
+    protected String getJSMainModuleName() {
+      return "index";
+    }
+  };
+
   @Override
   public ReactNativeHost getReactNativeHost() {
     return mReactNativeHost;
